[CFL] Compute Average Over TX Block Instead of Pred Block
When computing alpha, multiple averages are computed, one for each
transform block. The CfL prediction now uses the transform block average
instead of partition block average.

This allows the decoder to build the CfL prediction by using only the
collocated reconstructed luma values for the current transform size and
not the entire partition.

Results on Subset 1 (Compared to 9fcc095f1 with CfL)

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0413 |  0.2489 |  0.2606 |   0.0863 | 0.0430 |  0.0617 |     0.1555

Change-Id: Ib2ce2a3053b85300c5c62ef0e3270af489568a38