New experiment: LGT
In previous ADSTs, DST type 7 is used for length-4 and
DST type 4 is used for the other lenghts. In this LGT experiment
we explore other DSTs in between. When CONFIG_LGT flag is on,
adst4 and adst8 are replaced by lgt4 and lgt8, the intermediate
transforms with pre-tuned parameters.

The LGTs adopted here are lgt4_160 and lgt8_170, where the numbers
mean the self-loop weights times 100. This associated values in
DST type-7 and type-4 are 100 and 200.

ovr_psnr:
lowres: -0.163
midres: -0.063

These changes are not applied to the highbd scenario in the
current version.

Change-Id: I20600456da8766528b2b6b11aa28801e70af498e