Update patch set 9
Patch Set 9:

Yup, it basically makes a frame context per reference, and then uses the frame context from the reference in slot 0.

Additionally, before the encoder would use frame contexts based on the type of frame. Now it always uses LAST_FRAME, which causes the change in metrics. I would like to change it back to reorder the references so that the behavior is more like the old, but that'll be a separate patch.

Patch-set: 9
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>