[WIP] Inter and intra LGTs
Here we have an LGT to replace ADST in the intra scenarios, and another
LGT to replace ADST in inter scenarios. The changes are only applied to
transform length 4 and 8, and only for the lowbitdepth track.

In order to pass prediction information down to the transform level,
input arguments of some functions, such as av1_fht* and av1_iht*), are
changed.

Change-Id: Iadc1e02b53e3756b44f74ca648cfa8b0e8ca7af4