Inter and intra LGTs
Here we have an LGT to replace ADST for intra residual block, and
another LGT to replace ADST for inter residual blocks. The changes
are only applied to transform length 4 and 8, and only for the
lowbitdepth path.

Input arguments of some 2D transform functions, such as av1_fht* and
av1_iht*, are changed in order to support more adaptive transforms
in the future, where the LBT will be chosen based on predicted
pixels. Although LGT is applied to small transforms and LBD here, the
input arguments of larger transforms and HBD functions are changed
in the same way, just for consistency. Some tests are temporarily
disabled as a workaround, to avoid build failures.

Change-Id: Iadc1e02b53e3756b44f74ca648cfa8b0e8ca7af4