encoder: Remove 64x upsampled reference buffers
In addition to using massive amounts of memory and being a criminal
waste of cache (1 byte used for every 8 bytes fetched), they do not
handle border extension correctly (interpolation and border
extension do not commute unless you upsample into the border), nor
do they handle crop dimensions that are not a multiple of 8 (the
upsampled version is not sufficiently large).

This commit currently leaves in place the use_upsampled_references
speed feature, as it also controls whether to use 8-tap subpel
interpolation or bilinear subpel interpolation during searching,
and in some situations the amount of searching. This avoids a loss
in quality.

Upsampling on-demand instead of in advance appears to make the
encoder consistently faster by a small amount with --cpu-used=0.

Change-Id: I7527db9f83b87a7bb8b35342f7e6457cd0bef9cd