Update patch set 25
Patch Set 25:

> Patch Set 24: Code-Review+1
> 
> Okay for now. 
> However the coding of the references and whether the sign-bias bit is needed or not needs to be worked out subsequently.

Thanks for the suggestion. In the next patch it will add a to-do for the further exploration on the coding of the reference frames, a work to combine the binary tree and multiple symbol design.

The sign bias of the six reference frames cannot be fixed, for now. For the low delay cases, in the experiment of one-sided-compound, for instance, it will reload the BWDREF_FRAME and set it identical to LAST_FRAME, so that the compound reference pairs of (LAST,BWDREF), (LAST2,BWDREF), (LAST3,BWDREF), and (GOLDEN,BWDREF) will then indeed conduct the same compound reference prediction as (LAST,LAST),(LAST,LAST2), (LAST,LAST3), and (LAST,GOLDEN). In this case, BWDREF will then have the same sign bias as LAST_FRME. 

The tool of ext-comp-refs certainly is not in its final shape, as it needs to further combine with one-sided-compound to merge to one coding tool. I would suggest to have this coding tool firstly merged so that it would be convenient for the further exploration and stats collection.

Patch-set: 25
Reviewer: Gerrit User 5200 <5200@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>