Update patch set 1
Patch Set 1:

While I do think this patch should fix the mismatch bugs, the current behaviour still feels weird. Currently the code says:

* If frame-superres is disabled, we can use the mvs from cm->prev_frame if the size of cm matches the size of cm->prev_frame.

* If frame-superres is enabled, we can use the mvs from cm->prev_frame if the size of cm matches the size of *the latest encoded frame*.

For the current encoder, it looks like cm->prev_frame is either {latest encoded frame} or {LAST_FRAME, which is always set to the latest encoded+shown frame}. This is why this patch fixes the mismatches.

But, as far as I can tell, on the decoder side LAST_FRAME (and thus cm->prev_frame) could be set to any of the available reference buffers. So currently we're being too strict about when we can use the previous frame's mvs.

I think a better solution would be to always compare the sizes of cm and cm->prev_frame. We obviously can't use cm->prev_frame->buf.y_crop_width/height for this when frame-superres is enabled, but we could either compare the mi_rows/mi_cols values (which are already available), or add the (pre-superres) encoded size to RefCntBuffer.

Patch-set: 1
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>