Update patch set 1
Patch Set 1:

Having spoken to Deb and Zoe on the phone, it sounds like we want to change the prev_frame behavior further so that we can keep just one set of previous mvs (*), rather than the one set per ref frame we currently keep.

It's probably sensible to merge this for now, as a temporary fix, while I work on a patch to make the above change.

(*): Zoe mentioned that there is a separate experiment in development which will use >1 set of previous mvs. But for the baseline we only want one set.

Patch-set: 1
CC: Gerrit User 5200 <5200@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>