Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> It's weird that those values could be negative. Also there are multiple uses of such left shift, fixing only two of them is not enough. Joe, could you take a further look?

It can be negative if the linear interpolation involves the upper-left neighbor, which is at the -1 sample location (-2 if upsampled).

I'm okay with either the before or after versions.  Actually, I'm not sure why left shift of signed int is a problem, both versions should have identical resul t ?

Patch-set: 3