Update patch set 3
Patch Set 3:

> Actually, I'm not sure why left shift of signed int is a problem, both versions should have identical result ?

Short answer: because it's undefined behavior in C, and it might mislead the optimizer.

C doesn't make assumptions about the way signed values are represented by the hardware (as this could make it very hard to implement a C compiler for  architectures not using two's complement).

It becomes a problem when the compiler applies optimizations.

The optimizer will assume that your code won't trigger undefined behavior.
Thus, it will legitimately infer "a is always positive" if it's signed and left-shifted somewhere.

At this point, the optimizer might remove statements like "if(a < 0) { doSomething(); }", because it recognizes it as dead code.

Patch-set: 3