Update patch set 2
Patch Set 2:

Just a note that in general changing what gets built based on what happens to be installed on the build machine is a bad policy.

Imagine a Linux distro compiling packages for everyone, where maybe it has some of your build deps installed and maybe it doesn't, depending on the order it processed other packages, or which ones got assigned to that builder.

A better policy is if a dependency required for a default-enabled option is missing, then print a message telling the user how to disable that option. That ensures there's an explicit failure mode if someone fails to include a required pre-requisite, and if someone explicitly does not want a dependency on that pre-requisite, ensures they will explicitly disable the things that depend on it (so that those things don't randomly continue to get built if the undesired pre-requisite just happens to be installed).

Patch-set: 2
CC: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>