Update patch set 8
Patch Set 8:

> > I have results for a few videos: The BDRATE change (wrt. overall
 > PSNR) due to this patch is:
 > > tempete_cif.y4m: +0.046%
 > > akiyo_cif.y4m: -0.068%
 > > husky_cif.y4m: +0.234%
 > >
 > > This at least indicates that there isn't a huge quality drop.
 > 
 > I should add: These results are with my standard config for fast
 > tests, which is --disable-ext-refs --disable-ext-tx
 > --disable-rect-tx --disable-var-tx

I am not sure those results are representative of this patch as it currently stands.  Aren't the intra_ext_tx_cdf[] and inter_ext_tx_cdf[] arrays the ones that will not work with EC_ADAPT if you use a subset of the possible values?  That is part of the EXT_TX experiment which your testing disabled.

One thing I would suggest, is splitting out the changes to av1/encode/cost.[ch] into a separate patch.  That way we can update the costs one at a time to CDFs that we know will work with EC_ADAPT and the metric numbers will reflect the actual changes.

Patch-set: 8