Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > > I have results for a few videos: The BDRATE change (wrt. overall
>  > PSNR) due to this patch is:
>  > > tempete_cif.y4m: +0.046%
>  > > akiyo_cif.y4m: -0.068%
>  > > husky_cif.y4m: +0.234%
>  > >
>  > > This at least indicates that there isn't a huge quality drop.
>  > 
>  > I should add: These results are with my standard config for fast
>  > tests, which is --disable-ext-refs --disable-ext-tx
>  > --disable-rect-tx --disable-var-tx
> 
> I am not sure those results are representative of this patch as it currently stands.  Aren't the intra_ext_tx_cdf[] and inter_ext_tx_cdf[] arrays the ones that will not work with EC_ADAPT if you use a subset of the possible values?  That is part of the EXT_TX experiment which your testing disabled.
> 
> One thing I would suggest, is splitting out the changes to av1/encode/cost.[ch] into a separate patch.  That way we can update the costs one at a time to CDFs that we know will work with EC_ADAPT and the metric numbers will reflect the actual changes.

The reason I brought up making options in a CDF unavailable/giving them a probability of 0 was because we'd have to do that *if we want to get rid of the ind/inv arrays* (unless we come up with a better option). It's not what the current code does.

In the current code, for ext-tx, each context has its own CDF. The same number of elements in each CDF == the number of available transforms in that context. So as far as the entropy coder is concerned, these are just regular CDFs which can be adapted perfectly fine. The only bit of weirdness is that we have a context-dependent mapping between these smaller CDFs and a larger set of 16 values, but that's done outside the entropy coder.

IOW, I don't see any reason why this would have any problems with ext-tx, and indeed I've run a couple of single encodes and they work fine. I'll retest the three files with ext-tx enabled, but it'll take a little longer to get the results.

Patch-set: 8
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>