[CFL] Better encapsulation
The function cfl_compute_parameters is added and contains the logic
related to building the CfL context parameters. As such, many cfl
functions can now be encapsulated inside of cfl.c and not exposed to the
rest of AV1.

This also allows for supplemental asserts that validate that the CfL
context is properly built.

Change-Id: I6d14a426416b3af5491bdc145db7281b5e988cae