supertx: Read and write transforms with CDF in decode_partition
Patch b1bedf5f73 converted the three writes in bitstream.c that
specify an extended transform from using av1_write_token (encoded with
probability trees) to aom_write_symbol (encoded with CDFs).

That patch fixed up the two reads in decodemv.c but didn't fix up the
corresponding read in decodeframe.c. This patch does so.

The patch also fixes up a write of a (non-extended) transform when not
CONFIG_EXT_TX and the corresponding read.

Change-Id: Ibf5dcfcf3e7122f08dd0ef8616fb0ecddb95d99a