Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> if this were ever true in production you'd want to avoid this path; the assert won't do any good

I agree, the assert won't fire up in production.

It might fire up in development, though, if the buffer allocation/filling strategy is modified in an incompatible way.
And this filling strategy is, at the moment, the only thing that guarantees that this unrecoverable situation can't occur.

We're missing the forest for the trees here; the superframe index is tiny compared to the encoded frames, and 'av1_get_compressed_data', doesn't have an "output_buffer_size" parameter.

By the way, it's obvious that we're missing real error handling here ; however, it's not clear how the encoder should behave, should it detect that the encoded frame doesn't fit in its output buffer. Panic? Drop the frame? Re-encode with a higher quantizer?

Patch-set: 3