[CFL] Compute DC_PRED until end of frame
When computing CFL's prediction block level DC_PRED,
We can't assume that the neighbor will use the same
transform size. As such, we don't know what pixels
are available outside of the frame. A simple
solution is to stop computing DC_PRED at the block
boundary.

Results on Subset 1 ( Compared to 9c6f854 with CfL enabled)
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-0.0015 | -0.0026 | -0.0025 |  -0.0034 | -0.0008 | -0.0044 |     0.0019

Change-Id: I00ed4114e5c5871fdc7222fa2000474eb11d33e0