Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> It's intentional - because the reference frame slots have no other significance, we can use the ordering as signaling.
> 
> Note this patch is only half of what's needed - we also need the encoder to actually smartly choose the frame to go into the first slot (not just the actual last frame, like this patch currently does). I'll have a separate patch that implements the reordering in a similar manner to how it's currently done - a frame context matching the usage of the current frame (golden, altref, etc).

Understand the reordering of the frames in the reference buffer would always make it possible to have the frame context signalling frame always placed in slot 0. The concern here is that doing such adds another layer of constraint over the choice of slot 0. Not sure whether such constraint should be imposed, or we may still leave the freedom that the current frame context may choose any slot in the reference buffer through explicitly signalling the frame context index for the current coded frame.

Patch-set: 13