Update patch set 3
Patch Set 3:

I took a closer look at the issue, tracked down the un-aligned buffer to two temporary buffers on stack. 

Looking at the surrounding code, the issue was introduced when ext-inter was made default, which explains the time frame when the unit test started to show failures overnight. 

-Yaowu

Patch-set: 3