Add self-checker for CDF probabilities update
This is hidden behind a --enable-cdf-adapt-checking option. When the
option is set, this takes a copy of the CDFs that we ended up with at
the end of encode and compares that with those that we got at the end
of writing to the bitstream.

A slight complication is that a frame might have cm->interp_filter set
to SWITCHABLE so we update switchable_interp_cdf while
encoding. However, if we end up just choosing one filter, there's a
function called fix_interp_filter in bitstream.c which sensibly just
encodes that filter at top-level in the frame. If that happens, there
will be a spurious mismatch between switchable_interp_cdf's, which we
have special-case code to ignore.

Change-Id: I0cc0eec883ebce55d42c8ccf63450b6155236770