tempv_signaling: Simplify test for whether prev_frame works for mvs
For some background, see this previous change in Gerrit[0]. What's
going on here is that we only want to use a previous frame for motion
vector prediction if it's the same size as the current one. In this
context and when scaling with superres, "size" means "size before
scaling".

To check this correctly, we need to check prev_frame's width/height
and compare it with the current frame. Without superres, prev_frame's
width/height is stored in y_crop_width/y_crop_height so we can check
that way. With superres, those numbers are after the scaling, so can't
be compared with cm->width and cm->height.

The previous code worked around this by comparing with cm->last_width
and cm->last_height. That works because these are the width/height for
the last encoded and shown frame and that frame *is* prev_frame if
last_show_frame is true. Since this is the only case when we want to
use prev_frame, they are the numbers we need.

This patch simplifies the logic slightly by storing width/height in
RefCntBuffer before any scaling and checking that they match.

[0] https://aomedia-review.googlesource.com/c/13806

BUG=aomedia:78

Change-Id: If25227fa24222fc05c56529c2ac9ddf1e1c36a84