Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > On looking at things a bit more, my patch switches to costing from
>  > CDFs even when ec-adapt is disabled, whereas yours doesn't. So they
>  > aren't quite equivalent, but we need to decide one way or the
>  > other.
>  > 
>  > Both seem like valid choices, but I'd prefer to always use the CDFs
>  > for costing since (even without ec-adapt) it's ultimately the CDFs
>  > which decide how much space a symbol takes up in the bitstream.
> 
> Hi David, have a look at:
> 
> https://aomedia-review.googlesource.com/#/c/13860/
> 
> I've been working to remove the EC_ADAPT flag (that is have it enabled by default and thus remove all !EC_ADAPT code paths).  Because I have been asked to hold off landing 13860, I'd rather land 14447 even though it partially duplicates 13945 in the off chance that we are not able to remove EC_ADAPT.
> 
> I will update 13860 so that these new !EC_ADAPT paths that do not use the CDFs for costing are removed, and hopefully that patch will be approved and we will end up in the same place.

It's true that patch 13945 used to be based on top of 13860, but that isn't the case any more. So 13945 is also potentially mergeable at the moment.

If you remove the non-CDF costing, then the only differences between the patches seem to be:
i) You wrap the tree probability updates in #if !CONFIG_EC_ADAPT
ii) I cover a lot more cases

I feel it would be better to just have one patch which does both - ie, which removes all the things from 13945, but which also disables the tree probability updates if ec-adapt is enabled.

Patch-set: 4
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>