[CFL] clip CFL prediction to avoid overflow
The value predicted using CfL is clipped to avoid going out of the
scope of the uint8. Both overflow and underflow was detected over
Subtset1.

In cfl_alpha_dist, clamp is used because an int is desired to avoid
issues related to subtracting uint8 values.

Results on Subset1 (compared to 7e55571 with CfL enabled)

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0019 |  0.0001 |  0.0009 |   0.0047 | 0.0020 |  0.0023 |     0.0012

Change-Id: Ie1190e2286aa90542eaa68b814cc5cfa031acb73