Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Yeah, when bisecting for bugs 628 and 632 I hit 8b7a4e1 and noticed this oddity - I was curious if this would cause any of the Jenkins tests to fail.

It doesn't seem so, that's a coverage issue. Mostly for the encoder a unit test isn't quite right, we need a longer set of encodes to catch this kind of thing.

> But yeah, the same code is in all the other test files (maybe a revert of 8b7a4e1 and the linked issue would be better - or the removal of CONFIG_LOWBITDEPTH?)

let's do the revert to begin with, I really hadn't followed the CONFIG_LOWBITDEPTH thing, but I think the intent is to force 16-bit in all cases which to me is arguable and at the very least the implementation may not be correctly forcing all paths through the correct channels.

Patch-set: 1