Update patch set 1
Patch Set 1:

It sounds like that's more an argument to remove CONFIG_LOWBITDEPTH... I don't like aomdec and the unit tests doing different things.

My impression was that we wanted CONFIG_HIGHBITDEPTH on always (as the transforms are different), but 8-bit streams would in theory be encodeable and decodable with an 8-bit path. If this is wrong, and the intention is to really force a 12-bit path, we should be range expanding the input buffers as well.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>