Update patch set 1
Patch Set 1:

> With a lot of effort one could transition to a runtime config replacing all the compile time configurations. That has the benefit of at least keeping code paths from going stale from the build perspective.

Yes.

There's a middleground here: replacing some preprocessor symbols with global C constants.
("#if" directives are replaced with "if" statement in function bodies, and with a comment in struct definitions).

This doesn't require the effort of implementing runtime configuration, but already provides the benefit you're mentionning about the build.

We don't need to do it for all symbols to immediately reap some benefits.
(This would also allow running the unit tests of all experiments, regardless of the build configuration)

> Basing decisions based no the structure of VP9 probably isn't the best. The HBD path was a bit of an after thought that landed without much review.

I fully agree.

The way HBD was implemented in VP9 certainly isn't an example to follow, as it caused a lot of code duplication. It's also the root cause of the current situation about LBD/HBD.

The current LBD/HBD coding path duality started as an implementation detail from VP9, which, as you said, "landed without much review". This implementation detail, of course, doesn't appear in the VP9 specification pdf. It also doesn't need to appear in hardware implementations of VP9.

This is the part of VP9 I was referring to as "a good part".
We might be slowly loosing it with AV1.

Should we freeze the bitstream format to match the current state of the aomedia codebase, there would be a set of transforms for LBD, and another set of transforms for HBD. Both sets would need to appear in AV1 specification pdf and in hardware decoders. Worse: at the moment, we can't check if the LBD/HBD prediction functions match (potential mismatches are masked by transform mismatches).

By allowing the coding paths to diverge / stay in a diverged state, this implementation detail is slowly turning into a design goal.

Maybe that's what we want ; maybe we really do want, in the final specification, two different sets of transforms/prediction functions, one for each profile, realizing different computations.

This is an important design decision with many implications, and this decision should come from the people ; it shouldn't be the consequence of the failure to keep both coding paths in sync, especially if this failure is the consequence of an "after though without much review" from another codec.

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>