Update patch set 1
Patch Set 1:

So I think this patch isn't the actual fix we want.

I agree 2 different transforms is bad, but no one has stepped up to rewrite the LBD ones to match the HBD ones. We can just drop the LBD ones, but that would mean it'd be super easy to accidentally make a format that can't be decoded or encoded with 8 bit transform at all. As horrible has having 2 sets of transforms is, the latter is even worse.

What should we do for now? Ideally we'd quickly rewrite the LBD transforms, but no one has stepped up to do it, so we need a stopgap. My suggestion would be to revert the earlier two patches, and add CONFIG_LOWBITDEPTH to jenkins like an experiment (though it really isn't). Then we delete CONFIG_LOWBITDEPTH the moment fixed LBD transforms land. Would this be OK?

The other issue with this patch is that because lowbitepth is not a decoder default right now, API users must explicitly set it or fail to get a decoded frame. This means that the analyzer now has to bake in lowbitdepth as well...

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>