Update patch set 1
Patch Set 1:

> Could you please expand on this, ie. what would such a format look
 > like in practice?
 > (please keep in mind transforms don't have to directly work on
 > frame buffers, and thus can be mostly agnostic to the pixel format)

My main concern would be anything that truncates to 8-bit frame buffers - it's conceivable that some tool could slip in which could generate large reference buffer coefficients, which could then be predicted from. If the new transforms wrote to 8-bit buffers (even with a slow copy), or if we have sufficient range checks in place, then that'd be much less of a concern.

 > First, there's nothing more to test when CONFIG_LOWBITDEPTH changes
 > its value. This only changes the behaviour at main level, and could
 > as well be a command-line option.

Well, in theory all of its changes are covered by the transform/prediction unit tests, but the end to end tests are our last line of defense, and IMO should match what the user gets via ./configure. Also, the prediction tests don't seem to have 8-bit bounds checking, and there's also the matter of all the loop filters.

 > The full chain encode/decode tests in test_libaom aren't
 > explicitely targetting the transforms/prediction functions, and it
 > shouldn't be their role anyway. Ideally, the test result shouldn't
 > depend on the coding path used at all.
 > 
 > Moreover, transform/prediction functions are already unit-tested
 > separately, in a way that allows running the tests for both coding
 > paths, using only one build (where "everything" is enabled).
 > 
 > So, let's not multiply Jenkins builds, there already are lots of
 > them :-)

OK, so let's just pick one of the two to support for now. Jenkins tests should always match what the default configuration is though - the current tests always set 1 instead of CONFIG_LOWBITDEPTH for the end-to-end tests though, and I think that needs to change. We can leave the runtime control though.

 > We could instead use "decoder_cfg_t::allow_lowbitdepth" in a new
 > test (in test_libaom) encoding random extra-low resolution short
 > 8-bit noise sequences, decoding the bitstream with each coding
 > path, and comparing the YUV outputs.
 > Of course, such a test wouldn't pass until both coding paths are in
 > sync.

This is a very good idea, and an argument for keeping a runtime flag to switch between the two paths. Without any test, the chance that they drift apart again is basically 100%.

 > (actually, I'm beginning to
 > think that these new transforms should have been developped as an
 > experiment, maybe it's still possible to do so?).
Yeah... they were effectively part of the nextgenv2/misc_fixes baseline.

There are also the CONFIG_DAALA_DCT* experiments currently in Gerrit, these are transforms that both require smaller intermediates, and are identical in both bit depth paths. They are neither finished nor accepted though, so we need something to do in the meantime.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>