Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Here's the steps to reproduce the mismatch on patch set 2. It takes quite a long time to get to the issue.
> 
> make clean -s
> ../../configure --enable-internal-stats --enable-debug \
>   --disable-optimizations --disable-install-docs --log=yes \
>   --enable-av1 --enable-werror --enable-internal-stats \
>   --enable-experimental
> make -s -j 18
> 
> ./aomenc  ~/derf_cif/akiyo_cif.y4m -o /dev/null --codec=av1 --test-decode=fatal --good --cpu-used=0 --threads=0 --profile=0 --min-q=0 --max-q=63 --auto-alt-ref=1 --kf-max-dist=150 --kf-min-dist=0 --drop-frame=0 --static-thresh=0 --bias-pct=50 --minsection-pct=0 --maxsection-pct=2000 --arnr-maxframes=7 --arnr-strength=5 --sharpness=0 --undershoot-pct=100 --overshoot-pct=100 --tile-columns=0 --psnr --limit=50 --lag-in-frames=10 --resize-mode=2 --target-bitrate=100

Thanks for the catch. Re-reading my patch, it was never going to work - I was setting the width/height at the wrong time and so (in the example you send) a larger frame was trying to read motion vectors from a smaller one, which doesn't work very well...

I've just pushed a new version, which should get this right.

Patch-set: 4